individual mention command ./commands/mention.js
